+++ src/GlobalUsing.Infrastructure/InfrastructureJsonSerializerContext.cs
@@ -5,5 +5,6 @@ namespace GlobalUsing.Infrastructure;
 
 [JsonSourceGenerationOptions(WriteIndented = true)]
 [JsonSerializable(typeof(AnalysisResult))]
+[JsonSerializable(typeof(AnalysisSummary))]
 [JsonSerializable(typeof(ApplyResult))]
 internal sealed partial class InfrastructureJsonSerializerContext : JsonSerializerContext;